refactor: rewrite todolist-api, create auth-api, create types file, rewrite login/logout thunks to try catch & async await syntax, remove all then, remove error-utills, create JSdoc for utills functions, create common instance, fixed tests
