+++ src/api/instance.ts
@@ -1,0 +1,12 @@
+import axios from 'axios'
+
+const settings = {
+  withCredentials: true,
+  headers: {
+    'API-KEY': '1cdd9f77-c60e-4af5-b194-659e4ebd5d41',
+  },
+}
+export const instance = axios.create({
+  baseURL: 'https://social-network.samuraijs.com/api/1.1/',
+  ...settings,
+})
