+++ src/api/todolists-api.ts
@@ -1,18 +1,6 @@
-import axios from 'axios'
-import { UpdateDomainTaskModelType } from 'features/TodolistsList/tasks-slice'
+import { instance } from './instance'
+import { AddTaskArgs, GetTasksResponse, TaskType, UpdateTaskModelType } from './task-types'
 
-const settings = {
-  withCredentials: true,
-  headers: {
-    'API-KEY': '1cdd9f77-c60e-4af5-b194-659e4ebd5d41',
-  },
-}
-const instance = axios.create({
-  baseURL: 'https://social-network.samuraijs.com/api/1.1/',
-  ...settings,
-})
-
-// api
 export const todolistsAPI = {
   getTodolists() {
     const promise = instance.get<TodolistType[]>('todo-lists')
@@ -48,37 +36,11 @@ export const todolistsAPI = {
   },
 }
 
-export type AddTaskArgs = {
-  title: string
-  todolistId: string
-}
-export type UpdateTaskArgs = {
-  taskId: string
-  domainModel: UpdateDomainTaskModelType
-  todolistId: string
-}
-
 export type LoginParamsType = {
   email: string
   password: string
   rememberMe: boolean
   captcha?: string
-}
-
-export const authAPI = {
-  login(data: LoginParamsType) {
-    const promise = instance.post<ResponseType<{ userId?: number }>>('auth/login', data)
-    return promise
-  },
-  logout() {
-    const promise = instance.delete<ResponseType<{ userId?: number }>>('auth/login')
-    return promise
-  },
-  me() {
-    const promise =
-      instance.get<ResponseType<{ id: number; email: string; login: string }>>('auth/me')
-    return promise
-  },
 }
 
 // types
@@ -92,42 +54,4 @@ export type ResponseType<D = {}> = {
   resultCode: number
   messages: Array<string>
   data: D
-}
-export enum TaskStatuses {
-  New = 0,
-  InProgress = 1,
-  Completed = 2,
-  Draft = 3,
-}
-export enum TaskPriorities {
-  Low = 0,
-  Middle = 1,
-  Hi = 2,
-  Urgently = 3,
-  Later = 4,
-}
-export type TaskType = {
-  description: string
-  title: string
-  status: TaskStatuses
-  priority: TaskPriorities
-  startDate: string
-  deadline: string
-  id: string
-  todoListId: string
-  order: number
-  addedDate: string
-}
-export type UpdateTaskModelType = {
-  title: string
-  description: string
-  status: TaskStatuses
-  priority: TaskPriorities
-  startDate: string
-  deadline: string
-}
-type GetTasksResponse = {
-  error: string | null
-  totalCount: number
-  items: TaskType[]
 }
