+++ src/features/Login/Login.tsx
@@ -1,7 +1,7 @@
+/* eslint-disable react/jsx-no-target-blank */
 import React from 'react'
 import { useFormik } from 'formik'
 import { useSelector } from 'react-redux'
-import { loginTC } from './auth-slice'
 import { AppRootStateType } from '../../app/store'
 import { Navigate } from 'react-router-dom'
 import { useAppDispatch } from '../../hooks/useAppDispatch'
@@ -15,6 +15,7 @@ import {
   Grid,
   TextField,
 } from '@mui/material'
+import { authThunks } from './auth-slice'
 
 export const Login = () => {
   const dispatch = useAppDispatch()
@@ -40,7 +41,7 @@ export const Login = () => {
       rememberMe: false,
     },
     onSubmit: (values) => {
-      dispatch(loginTC(values))
+      dispatch(authThunks.login(values))
     },
   })
