+++ src/features/TodolistsList/Todolist/Todolist.tsx
@@ -3,12 +3,13 @@ import React, { useCallback, useEffect } from 'react'
 import { AddItemForm } from '../../../components/AddItemForm/AddItemForm'
 import { EditableSpan } from '../../../components/EditableSpan/EditableSpan'
 import { Task } from './Task/Task'
-import { TaskStatuses, TaskType } from '../../../api/todolists-api'
 import { FilterValuesType, TodolistDomainType } from '../todolists-slice'
 import { tasksThunks } from '../tasks-slice'
 import { useAppDispatch } from '../../../hooks/useAppDispatch'
 import { Button, IconButton } from '@mui/material'
 import { Delete } from '@mui/icons-material'
+import { TaskType } from 'api/task-types'
+import { TaskStatuses } from 'utils/enums'
 
 type PropsType = {
   todolist: TodolistDomainType
