+++ src/utils/enums.ts
@@ -1,0 +1,19 @@
+export const ResultCode = {
+  success: 0,
+  error: 1,
+  captcha: 10,
+} as const
+
+export enum TaskStatuses {
+  New = 0,
+  InProgress = 1,
+  Completed = 2,
+  Draft = 3,
+}
+export enum TaskPriorities {
+  Low = 0,
+  Middle = 1,
+  Hi = 2,
+  Urgently = 3,
+  Later = 4,
+}
